\docType{package}
\name{cdtRefficient-package}
\alias{cdtRefficient}
\alias{cdtRefficient-package}
\title{R courses for Newcastle University's CDT in Cloud Computing for Big Data}
\description{
  Functions and datasets used for Newcastle University's 
  CDT in Cloud Computing for Big Data. This package is 
  used in the "Efficient R Programming" course.
}
\keyword{package}
